Cleanup library and moved things to better locations
